Move config handling out of state.rs
