+++ src/config.rs
@@ -1,12 +1,22 @@
 pub mod api;
 
-use std::path::Path;
+use crate::config::api::{msg::ModifierMask, PinnacleSocketSource};
+use std::{
+    path::{Path, PathBuf},
+    sync::{Arc, Mutex},
+};
 
 use anyhow::Context;
+use calloop::channel::Sender;
 use smithay::input::keyboard::keysyms;
 use toml::Table;
 
 use api::msg::Modifier;
+
+use crate::{
+    state::{State, WithState},
+    tag::TagId,
+};
 
 #[derive(serde::Deserialize, Debug)]
 pub struct Metaconfig {
@@ -97,11 +107,169 @@ pub enum Key {
     Escape = keysyms::KEY_Escape,
 }
 
-pub fn parse(config_dir: &Path) -> anyhow::Result<Metaconfig> {
+fn parse(config_dir: &Path) -> anyhow::Result<Metaconfig> {
     let config_dir = config_dir.join("metaconfig.toml");
 
     let metaconfig =
         std::fs::read_to_string(config_dir).context("Failed to read metaconfig.toml")?;
 
     toml::from_str(&metaconfig).context("Failed to deserialize toml")
+}
+
+fn get_config_dir() -> PathBuf {
+    let config_dir = std::env::var("PINNACLE_CONFIG_DIR")
+        .ok()
+        .and_then(|s| Some(PathBuf::from(shellexpand::full(&s).ok()?.to_string())));
+
+    config_dir.unwrap_or(crate::XDG_BASE_DIRS.get_config_home())
+}
+
+pub fn start_config(tx_channel: Sender<api::msg::Msg>) -> anyhow::Result<ConfigReturn> {
+    let config_dir = get_config_dir();
+    tracing::debug!("config dir is {:?}", config_dir);
+
+    let metaconfig = parse(&config_dir)?;
+
+    // If a socket is provided in the metaconfig, use it.
+    let socket_dir = if let Some(socket_dir) = &metaconfig.socket_dir {
+        // cd into the metaconfig dir and canonicalize to preserve relative paths
+        // like ./dir/here
+        let current_dir = std::env::current_dir()?;
+
+        std::env::set_current_dir(&config_dir)?;
+        let socket_dir = PathBuf::from(socket_dir).canonicalize()?;
+        std::env::set_current_dir(current_dir)?;
+        socket_dir
+    } else {
+        // Otherwise, use $XDG_RUNTIME_DIR. If that doesn't exist, use /tmp.
+        crate::XDG_BASE_DIRS
+            .get_runtime_directory()
+            .cloned()
+            .unwrap_or(PathBuf::from(crate::config::api::DEFAULT_SOCKET_DIR))
+    };
+
+    let socket_source = PinnacleSocketSource::new(tx_channel, &socket_dir)
+        .context("Failed to create socket source")?;
+
+    let reload_keybind = metaconfig.reload_keybind;
+    let kill_keybind = metaconfig.kill_keybind;
+
+    let mut command = metaconfig.command.split(' ');
+
+    let arg1 = command
+        .next()
+        .context("command in metaconfig.toml was empty")?;
+
+    std::env::set_var("PINNACLE_DIR", std::env::current_dir()?);
+
+    let envs = metaconfig
+        .envs
+        .unwrap_or(toml::map::Map::new())
+        .into_iter()
+        .filter_map(|(key, val)| {
+            if let toml::Value::String(string) = val {
+                Some((
+                    key,
+                    shellexpand::full_with_context(
+                        &string,
+                        || std::env::var("HOME").ok(),
+                        // Expand nonexistent vars to an empty string instead of crashing
+                        |var| Ok::<_, ()>(Some(std::env::var(var).unwrap_or("".to_string()))),
+                    )
+                    .ok()?
+                    .to_string(),
+                ))
+            } else {
+                None
+            }
+        })
+        .collect::<Vec<_>>();
+
+    tracing::debug!("Config envs are {:?}", envs);
+
+    // Using async_process's Child instead of std::process because I don't have to spawn my own
+    // thread to wait for the child
+    let child = async_process::Command::new(arg1)
+        .args(command)
+        .envs(envs)
+        .current_dir(config_dir)
+        .stdout(async_process::Stdio::inherit())
+        .stderr(async_process::Stdio::inherit())
+        .spawn()
+        .expect("failed to spawn config");
+
+    tracing::info!("Started config with {}", metaconfig.command);
+
+    let reload_mask = ModifierMask::from(reload_keybind.modifiers);
+    let kill_mask = ModifierMask::from(kill_keybind.modifiers);
+
+    Ok(ConfigReturn {
+        reload_keybind: (reload_mask, reload_keybind.key as u32),
+        kill_keybind: (kill_mask, kill_keybind.key as u32),
+        config_child_handle: child,
+        socket_source,
+    })
+}
+
+pub struct ConfigReturn {
+    pub reload_keybind: (ModifierMask, u32),
+    pub kill_keybind: (ModifierMask, u32),
+    pub config_child_handle: async_process::Child,
+    pub socket_source: PinnacleSocketSource,
+}
+
+impl State {
+    pub fn restart_config(&mut self) -> anyhow::Result<()> {
+        tracing::info!("Restarting config");
+        tracing::debug!("Clearing tags");
+
+        for output in self.space.outputs() {
+            output.with_state(|state| state.tags.clear());
+        }
+
+        TagId::reset();
+
+        tracing::debug!("Clearing mouse and keybinds");
+        self.input_state.keybinds.clear();
+        self.input_state.mousebinds.clear();
+        self.window_rules.clear();
+
+        tracing::debug!("Killing old config");
+        if let Err(err) = self.api_state.config_process.kill() {
+            tracing::warn!("Error when killing old config: {err}");
+        }
+
+        self.loop_handle.remove(self.api_state.socket_token);
+
+        let ConfigReturn {
+            reload_keybind,
+            kill_keybind,
+            config_child_handle,
+            socket_source,
+        } = start_config(self.api_state.tx_channel.clone())?;
+
+        let socket_token = self
+            .loop_handle
+            .insert_source(socket_source, |stream, _, data| {
+                if let Some(old_stream) = data
+                    .state
+                    .api_state
+                    .stream
+                    .replace(Arc::new(Mutex::new(stream)))
+                {
+                    old_stream
+                        .lock()
+                        .expect("Couldn't lock old stream")
+                        .shutdown(std::net::Shutdown::Both)
+                        .expect("Couldn't shutdown old stream");
+                }
+            })?;
+
+        self.input_state.reload_keybind = reload_keybind;
+        self.input_state.kill_keybind = kill_keybind;
+        self.api_state.config_process = config_child_handle;
+        self.api_state.socket_token = socket_token;
+
+        Ok(())
+    }
 }
